LayoutTests/storage/indexeddb/modern/autoincrement-abort-expected.txt

This test creates a new database with an objectstore that autoincrements. It then puts some things in that object store, checking the keys that were used. But it then aborts that transaction. Then it opens a new one and puts something in it, double checking that the key generator was reverted when the above transaction was aborted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Initial upgrade versionchange transaction complete
ALERT: Opening readwrite transaction to bump the key generator, but it will be aborted
ALERT: Key used for put was 1
ALERT: Key used for put was 2
ALERT: Key used for put was 3
ALERT: Key used for put was 4
ALERT: Key used for put was 5
ALERT: Key used for put was 6
ALERT: readwrite transaction abort
ALERT: Opening readwrite transaction to make sure the key generator had successfully reverted
ALERT: Key used for put was 1
ALERT: Key used for put was 2
ALERT: Key used for put was 3
ALERT: Key used for put was 4
ALERT: Key used for put was 5
ALERT: Key used for put was 6
ALERT: readwrite transaction complete
ALERT: Opening readwrite transaction to make sure the key generator picks up where it should've left off
ALERT: Key used for put was 7
ALERT: Key used for put was 8
ALERT: Key used for put was 9
ALERT: Key used for put was 10
ALERT: Key used for put was 11
ALERT: Key used for put was 12
ALERT: readwrite transaction complete
PASS successfullyParsed is true

TEST COMPLETE
